+++ server/src-lib/Hasura/Backends/DataConnector/Adapter/Metadata.hs
@@ -9,7 +9,7 @@ import Data.Environment (Environment)
 import Data.HashMap.Strict qualified as Map
 import Data.HashMap.Strict.InsOrd qualified as OMap
 import Data.Sequence qualified as Seq
-import Data.Sequence.NonEmpty qualified as NESeq
+import Data.Sequence.NESeq qualified as NESeq
 import Data.Text qualified as Text
 import Data.Text.Extended (toTxt, (<<>), (<>>))
 import Hasura.Backends.DataConnector.API qualified as API