+++ server/src-test/Hasura/QuickCheck/Instances.hs
@@ -361,6 +361,7 @@ instance Arbitrary TableCustomRootFields where
         <*> arbitrary
         <*> arbitrary
         <*> arbitrary
+        <*> arbitrary
     )
       `suchThat` allFieldNamesAreUnique
     where